% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identifySA}
\alias{identifySA}
\title{Check identifiability conditions}
\usage{
identifySA(lamcon)
}
\arguments{
\item{lamcon}{Matrix of number of sources by number of constituents containing the 0s and 1s for identifiability conditions}
}
\description{
\code{identifySA} checks a identifiability conditions for source apportionment
}
\details{
This function checks conditions C1- C3-1 from Park et al. 2002 Environmetrics
}

