% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcmcsa}
\alias{mcmcsa}
\title{MCMC for source apportionment}
\usage{
mcmcsa(dat, L, lamcon, mdls = NULL, guessvec = NULL, burnin = 10000,
  N = 1e+05)
}
\arguments{
\item{data}{data frame of daily constituent concentrations with date as first column}

\item{L}{number of sources}

\item{lamcon}{Matrix of sources by constituents with 1 and 0 for conditions}

\item{mdls}{data frame of daily MDL values of the same dimensions of data}

\item{guessvec}{starting values (do not need to provide)}

\item{burnin}{number of burnin samples}

\item{N}{number of samples}
}
\description{
\code{mcmcsa} performs MCMC for source apportionment
}
\details{
This function aims to draw from the posterior distributions of the source
profiles, the source contributions/concentrations, the variance of the data
and the other parameters.  Eventually, this function will incorporate
options for concentrations observed below the MDL
}
\examples{
data(nycdat)
data(nycmdl)
#fix data totals
pm <- nycdat$PM25
whPM <- which(colnames(nycdat) == "PM25")
nycdat <- nycdat[, -whPM]
whPM <- which(colnames(nycmdl) == "PM25")
nycmdl <- nycmdl[, -whPM]
mcmcsa(nycdat, nycmdl)
}

