% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_outer.R
\name{mcmcsa}
\alias{mcmcsa}
\alias{mcmcsa.default}
\title{MCMC for source apportionment}
\usage{
mcmcsa(x, ...)

\method{mcmcsa}{default}(dat, lamcon, mdls = NULL, guessvec = NULL,
  burnin = 10000, N = 1e+05, fix = NULL)
}
\arguments{
\item{dat}{data frame of daily constituent concentrations with date as first column}

\item{lamcon}{Matrix of sources by constituents with 1 and 0 for conditions}

\item{mdls}{data frame of daily MDL values of the same dimensions of data}

\item{guessvec}{starting values (do not need to provide)}

\item{burnin}{number of burnin samples}

\item{N}{number of samples}

\item{L}{number of sources}
}
\description{
\code{mcmcsa} performs MCMC for source apportionment
}
\details{
This function aims to draw from the posterior distributions of the source
profiles, the source contributions/concentrations, the variance of the data
and the other parameters.  Eventually, this function will incorporate 
options for concentrations observed below the MDL
}
\examples{
data(nycdat)
data(nycmdl)
data(lamcon)
#fix data totals
pm <- nycdat$PM25
whPM <- which(colnames(nycdat) == "PM25")
nycdat <- nycdat[, -whPM]
whPM <- which(colnames(nycmdl) == "PM25")
nycmdl <- nycmdl[, -whPM]
L <- 4

mcmcsa(nycdat, L, lamcon)
}

